Create codechef1
Chef and Chefina are playing with dice. In one turn, both of them roll their dice at once.

They consider a turn to be good if the sum of the numbers on their dice is greater than 
6
6.
Given that in a particular turn Chef and Chefina got 
�
X and 
�
Y on their respective dice, find whether the turn was good.

Input Format
The first line of input will contain a single integer 
�
T, denoting the number of test cases.
Each test case contains two space-separated integers 
�
X and 
�
Y — the numbers Chef and Chefina got on their respective dice.